feat: Move tool definitions to /lib/tools.ts
